// cgocore-ui/src/app/features/models/schedule/views/basic/DatedFlightBasicView.ts
/**
 * auto generated basic view bean ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightBasicView {

id: number;
selected: boolean;
canDelete: boolean;
companyIataCode: string;
flightNumber: string;
departureDate: string;
originIataCode: string;
departureTime: string;
destinationIataCode: string;
arrivalDate: string;
arrivalTime: string;

/* Specific Code Start */
/* Specific Code End */
}