// cgocore-ui/src/app/features/models/schedule/filters/DatedFlightFilter.ts
/**
 * auto generated filter ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightFilter {

companyIataCode: string;
flightNumber: string;
departureDateMinValue: string;
departureDateMaxValue: string;
originIataCode: string;
destinationIataCode: string;

/* Specific Code Start */
/* Specific Code End */
}