// cgocore-ui/src/app/features/models/schedule/sortings/DatedFlightSorting.ts
/**
 * auto generated sorting ts file
 * <br/>write modifications between specific code marks
 * <br/>processed by skeleton-generator
 */
export class DatedFlightSorting {

companyIataCodeOrderType: string;
flightNumberOrderType: string;
departureDateOrderType: string;
originIataCodeOrderType: string;
departureTimeOrderType: string;
destinationIataCodeOrderType: string;
arrivalDateOrderType: string;
arrivalTimeOrderType: string;

/* Specific Code Start */
/* Specific Code End */
}